test(x/ecocredit): marketplace update acceptance tests
