+++ x/ecocredit/server/marketplace/msg_update_sell_orders_test.go
@@ -1,251 +1,513 @@
 package marketplace
 
 import (
+	"strconv"
 	"testing"
 	"time"
 
-	"gotest.tools/v3/assert"
-
 	"github.com/cosmos/cosmos-sdk/orm/types/ormerrors"
-	"github.com/cosmos/cosmos-sdk/testutil/testdata"
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
-
-	marketApi "github.com/regen-network/regen-ledger/api/regen/ecocredit/marketplace/v1"
+	"github.com/gogo/protobuf/jsonpb"
+	"github.com/regen-network/gocuke"
+	"github.com/regen-network/regen-ledger/types"
 	"github.com/regen-network/regen-ledger/types/math"
-	"github.com/regen-network/regen-ledger/x/ecocredit"
+	"github.com/stretchr/testify/require"
+	"google.golang.org/protobuf/types/known/timestamppb"
+
+	api "github.com/regen-network/regen-ledger/api/regen/ecocredit/marketplace/v1"
+	coreapi "github.com/regen-network/regen-ledger/api/regen/ecocredit/v1"
 	"github.com/regen-network/regen-ledger/x/ecocredit/marketplace"
 )
 
-func TestUpdateSellOrders_QuantityAndAutoRetire(t *testing.T) {
-	t.Parallel()
-	s := setupBase(t)
-	s.testSellSetup(batchDenom, ask.Denom, ask.Denom[1:], classId, start, end, creditType)
-
-	expiration := time.Now()
-	_, err := s.k.Sell(s.ctx, &marketplace.MsgSell{
-		Owner: s.addr.String(),
-		Orders: []*marketplace.MsgSell_Order{
-			{BatchDenom: batchDenom, Quantity: "5.22", AskPrice: &ask, DisableAutoRetire: false, Expiration: &expiration},
-			{BatchDenom: batchDenom, Quantity: "30", AskPrice: &ask, DisableAutoRetire: true, Expiration: &expiration},
-		},
-	})
-	assert.NilError(t, err)
-
-	balBefore, _ := s.getBalanceAndSupply(1, s.addr)
-
-	_, err = s.k.UpdateSellOrders(s.ctx, &marketplace.MsgUpdateSellOrders{
-		Owner: s.addr.String(),
-		Updates: []*marketplace.MsgUpdateSellOrders_Update{
-			{SellOrderId: 1, NewQuantity: "10", DisableAutoRetire: true},
-			{SellOrderId: 2, NewQuantity: "28.7232", DisableAutoRetire: false},
-		},
-	})
-	assert.NilError(t, err)
-
-	balAfter, _ := s.getBalanceAndSupply(1, s.addr)
-
-	// sellOrder 1: 5.22 originally, increased by 10 = change of 4.78
-	// sellOrder 2: 30 originally, decreased by 28.7232 = change of -1.2768
-	// total change: 4.78 + (-1.2768) = 3.5032
-
-	actualEscrowChange, err := math.NewDecFromString("3.5032")
-	assert.NilError(t, err)
-
-	assertCreditsEscrowed(t, balBefore, balAfter, actualEscrowChange)
-
-	order1, err := s.marketStore.SellOrderTable().Get(s.ctx, 1)
-	assert.NilError(t, err)
-	order2, err := s.marketStore.SellOrderTable().Get(s.ctx, 2)
-	assert.NilError(t, err)
-	assert.Equal(t, "10", order1.Quantity)
-	assert.Equal(t, "28.7232", order2.Quantity)
-
-	assert.Equal(t, true, order1.DisableAutoRetire)
-	assert.Equal(t, false, order2.DisableAutoRetire)
+type updateSellOrdersSuite struct {
+	*baseSuite
+	alice               sdk.AccAddress
+	bob                 sdk.AccAddress
+	aliceTradableAmount string
+	aliceEscrowedAmount string
+	creditTypeAbbrev    string
+	classId             string
+	batchDenom          string
+	sellOrderId         uint64
+	askPrice            *sdk.Coin
+	quantity            string
+	disableAutoRetire   bool
+	expiration          *time.Time
+	res                 *marketplace.MsgUpdateSellOrdersResponse
+	err                 error
 }
 
-func TestUpdateSellOrders_QuantityInvalid(t *testing.T) {
-	t.Parallel()
-	s := setupBase(t)
-	s.testSellSetup(batchDenom, ask.Denom, ask.Denom[1:], classId, start, end, creditType)
-
-	expiration := time.Now()
-	_, err := s.k.Sell(s.ctx, &marketplace.MsgSell{
-		Owner: s.addr.String(),
-		Orders: []*marketplace.MsgSell_Order{
-			{BatchDenom: batchDenom, Quantity: "5.22", AskPrice: &ask, DisableAutoRetire: false, Expiration: &expiration},
-			{BatchDenom: batchDenom, Quantity: "30", AskPrice: &ask, DisableAutoRetire: true, Expiration: &expiration},
-		},
-	})
-	assert.NilError(t, err)
-
-	// cannot update sell order that does not exist
-	_, err = s.k.UpdateSellOrders(s.ctx, &marketplace.MsgUpdateSellOrders{
-		Owner: s.addr.String(),
-		Updates: []*marketplace.MsgUpdateSellOrders_Update{
-			{SellOrderId: 25, NewQuantity: "3"},
-		},
-	})
-	assert.ErrorContains(t, err, ormerrors.NotFound.Error())
-
-	// cannot increase sell order with more credits than in balance
-	_, err = s.k.UpdateSellOrders(s.ctx, &marketplace.MsgUpdateSellOrders{
-		Owner: s.addr.String(),
-		Updates: []*marketplace.MsgUpdateSellOrders_Update{
-			{SellOrderId: 1, NewQuantity: "1000000000"},
-		},
-	})
-	assert.ErrorContains(t, err, ecocredit.ErrInsufficientCredits.Error())
-
-	// cannot increase sell order with higher precision than credit type
-	_, err = s.k.UpdateSellOrders(s.ctx, &marketplace.MsgUpdateSellOrders{
-		Owner: s.addr.String(),
-		Updates: []*marketplace.MsgUpdateSellOrders_Update{
-			{SellOrderId: 1, NewQuantity: "10.329083409234908234"},
-		},
-	})
-	assert.ErrorContains(t, err, "exceeds maximum decimal places")
+func TestUpdateSellOrders(t *testing.T) {
+	gocuke.NewRunner(t, &updateSellOrdersSuite{}).Path("./features/msg_update_sell_orders.feature").Run()
 }
 
-func TestUpdateSellOrders_Unauthorized(t *testing.T) {
-	t.Parallel()
-	s := setupBase(t)
-	s.testSellSetup(batchDenom, ask.Denom, ask.Denom[1:], classId, start, end, creditType)
-	_, _, unauthorized := testdata.KeyTestPubAddr()
-	expiration := time.Now()
-	_, err := s.k.Sell(s.ctx, &marketplace.MsgSell{
-		Owner: s.addr.String(),
-		Orders: []*marketplace.MsgSell_Order{
-			{BatchDenom: batchDenom, Quantity: "5.22", AskPrice: &ask, DisableAutoRetire: false, Expiration: &expiration},
-			{BatchDenom: batchDenom, Quantity: "30", AskPrice: &ask, DisableAutoRetire: true, Expiration: &expiration},
-		},
-	})
-	assert.NilError(t, err)
-
-	// cannot edit the sell order with this address
-	_, err = s.k.UpdateSellOrders(s.ctx, &marketplace.MsgUpdateSellOrders{
-		Owner: unauthorized.String(),
-		Updates: []*marketplace.MsgUpdateSellOrders_Update{
-			{SellOrderId: 1, NewQuantity: "1"},
-		},
-	})
-	assert.ErrorContains(t, err, sdkerrors.ErrUnauthorized.Error())
+func (s *updateSellOrdersSuite) Before(t gocuke.TestingT) {
+	s.baseSuite = setupBase(t)
+	s.alice = s.addr
+	s.bob = s.addr2
+	s.aliceTradableAmount = "200"
+	s.creditTypeAbbrev = "C"
+	s.classId = "C01"
+	s.batchDenom = "C01-001-20200101-20210101-001"
+	s.sellOrderId = 1
+	s.askPrice = &sdk.Coin{
+		Denom:  "regen",
+		Amount: sdk.NewInt(100),
+	}
+	s.quantity = "100"
 }
 
-func TestUpdateSellOrder_AskPrice(t *testing.T) {
-	t.Parallel()
-	s := setupBase(t)
-	s.testSellSetup(batchDenom, ask.Denom, ask.Denom[1:], classId, start, end, creditType)
+func (s *updateSellOrdersSuite) ABlockTimeWithTimestamp(a string) {
+	blockTime, err := types.ParseDate("block time", a)
+	require.NoError(s.t, err)
 
-	expiration := time.Now()
-	_, err := s.k.Sell(s.ctx, &marketplace.MsgSell{
-		Owner: s.addr.String(),
-		Orders: []*marketplace.MsgSell_Order{
-			{BatchDenom: batchDenom, Quantity: "5.22", AskPrice: &ask, DisableAutoRetire: false, Expiration: &expiration},
-		},
-	})
-	assert.NilError(t, err)
-
-	orderBefore, err := s.marketStore.SellOrderTable().Get(s.ctx, 1)
-	// can update price of same denom
-	askUpdate := sdk.NewInt64Coin(ask.Denom, 25)
-	_, err = s.k.UpdateSellOrders(s.ctx, &marketplace.MsgUpdateSellOrders{
-		Owner: s.addr.String(),
-		Updates: []*marketplace.MsgUpdateSellOrders_Update{
-			{SellOrderId: 1, NewAskPrice: &askUpdate},
-		},
-	})
-	assert.NilError(t, err)
-	orderAfter, err := s.marketStore.SellOrderTable().Get(s.ctx, 1)
-	assert.NilError(t, err)
-	assert.Equal(t, orderAfter.AskAmount, askUpdate.Amount.String())
-	assert.Equal(t, orderBefore.MarketId, orderAfter.MarketId)
-
-	orderBefore, err = s.marketStore.SellOrderTable().Get(s.ctx, 1)
-	assert.NilError(t, err)
-
-	// can update price with new denom in allowed ask denoms
-	askUpdate = sdk.NewInt64Coin("ubar", 18)
-	assert.NilError(t, s.marketStore.AllowedDenomTable().Insert(s.ctx, &marketApi.AllowedDenom{
-		BankDenom:    askUpdate.Denom,
-		DisplayDenom: askUpdate.Denom,
-	}))
-	_, err = s.k.UpdateSellOrders(s.ctx, &marketplace.MsgUpdateSellOrders{
-		Owner: s.addr.String(),
-		Updates: []*marketplace.MsgUpdateSellOrders_Update{
-			{SellOrderId: 1, NewAskPrice: &askUpdate},
-		},
-	})
-	assert.NilError(t, err)
-
-	order, err := s.marketStore.SellOrderTable().Get(s.ctx, 1)
-	assert.NilError(t, err)
-
-	assert.Equal(t, order.AskAmount, askUpdate.Amount.String())
-	assert.Equal(t, order.MarketId, orderBefore.MarketId+1)
-}
-
-func TestUpdateSellOrder_Expiration(t *testing.T) {
-	t.Parallel()
-	s := setupBase(t)
-	s.testSellSetup(batchDenom, ask.Denom, ask.Denom[1:], classId, start, end, creditType)
-
-	future := time.Date(2077, 1, 1, 1, 1, 1, 1, time.Local)
-	middle := time.Date(2022, 1, 1, 1, 1, 1, 1, time.Local)
-	past := time.Date(1970, 1, 1, 1, 1, 1, 1, time.Local)
-
-	// create a sell order - expiration does not matter at this point, as the block time has not been set yet.
-	_, err := s.k.Sell(s.ctx, &marketplace.MsgSell{
-		Owner: s.addr.String(),
-		Orders: []*marketplace.MsgSell_Order{
-			{BatchDenom: batchDenom, Quantity: "5.22", AskPrice: &ask, DisableAutoRetire: false, Expiration: &past},
-		},
-	})
-	assert.NilError(t, err)
-
-	s.sdkCtx = s.sdkCtx.WithBlockTime(middle)
+	s.sdkCtx = s.sdkCtx.WithBlockTime(blockTime)
 	s.ctx = sdk.WrapSDKContext(s.sdkCtx)
-
-	// should work with future time
-	_, err = s.k.UpdateSellOrders(s.ctx, &marketplace.MsgUpdateSellOrders{
-		Owner: s.addr.String(),
-		Updates: []*marketplace.MsgUpdateSellOrders_Update{
-			{SellOrderId: 1, NewExpiration: &future},
-		},
-	})
-	assert.NilError(t, err)
-
-	// should work with the same time as blockTime
-	_, err = s.k.UpdateSellOrders(s.ctx, &marketplace.MsgUpdateSellOrders{
-		Owner: s.addr.String(),
-		Updates: []*marketplace.MsgUpdateSellOrders_Update{
-			{SellOrderId: 1, NewExpiration: &middle},
-		},
-	})
-	assert.NilError(t, err)
-
-	// should not work with past time
-	_, err = s.k.UpdateSellOrders(s.ctx, &marketplace.MsgUpdateSellOrders{
-		Owner: s.addr.String(),
-		Updates: []*marketplace.MsgUpdateSellOrders_Update{
-			{SellOrderId: 1, NewExpiration: &past},
-		},
-	})
-	assert.ErrorContains(t, err, "expiration must be in the future")
 }
 
-func TestSellOrder_InvalidDenom(t *testing.T) {
-	t.Parallel()
-	s := setupBase(t)
-	s.testSellSetup(batchDenom, ask.Denom, ask.Denom[1:], classId, start, end, creditType)
-	invalidAsk := sdk.NewInt64Coin("ubar", 10)
-	expiration := time.Now()
-	_, err := s.k.Sell(s.ctx, &marketplace.MsgSell{
-		Owner: s.addr.String(),
-		Orders: []*marketplace.MsgSell_Order{
-			{BatchDenom: batchDenom, Quantity: "5.22", AskPrice: &invalidAsk, DisableAutoRetire: false, Expiration: &expiration},
+func (s *updateSellOrdersSuite) ACreditType() {
+	err := s.coreStore.CreditTypeTable().Insert(s.ctx, &coreapi.CreditType{
+		Abbreviation: s.creditTypeAbbrev,
+	})
+	require.NoError(s.t, err)
+}
+
+func (s *updateSellOrdersSuite) ACreditTypeWithAbbreviation(a string) {
+	err := s.coreStore.CreditTypeTable().Insert(s.ctx, &coreapi.CreditType{
+		Abbreviation: a,
+	})
+	require.NoError(s.t, err)
+}
+
+func (s *updateSellOrdersSuite) ACreditTypeWithPrecision(a string) {
+	precision, err := strconv.ParseUint(a, 10, 32)
+	require.NoError(s.t, err)
+
+	err = s.coreStore.CreditTypeTable().Insert(s.ctx, &coreapi.CreditType{
+		Abbreviation: s.creditTypeAbbrev,
+		Precision:    uint32(precision),
+	})
+	require.NoError(s.t, err)
+}
+
+func (s *updateSellOrdersSuite) AnAllowedDenom() {
+	err := s.marketStore.AllowedDenomTable().Insert(s.ctx, &api.AllowedDenom{
+		BankDenom: s.askPrice.Denom,
+	})
+	require.NoError(s.t, err)
+}
+
+func (s *updateSellOrdersSuite) AnAllowedDenomWithBankDenom(a string) {
+	err := s.marketStore.AllowedDenomTable().Insert(s.ctx, &api.AllowedDenom{
+		BankDenom:    a,
+		DisplayDenom: a,
+	})
+	require.NoError(s.t, err)
+}
+
+func (s *updateSellOrdersSuite) AMarketWithCreditTypeAndBankDenom(a string, b string) {
+	s.marketStore.MarketTable().Insert(s.ctx, &api.Market{
+		CreditTypeAbbrev: a,
+		BankDenom:        b,
+	})
+}
+
+func (s *updateSellOrdersSuite) AliceCreatedASellOrder() {
+	s.sellOrderSetup(1)
+}
+
+func (s *updateSellOrdersSuite) AliceCreatedASellOrderWithId(a string) {
+	id, err := strconv.ParseUint(a, 10, 32)
+	require.NoError(s.t, err)
+
+	s.sellOrderId = id
+
+	s.sellOrderSetup(1)
+}
+
+func (s *updateSellOrdersSuite) AliceCreatedASellOrderWithQuantity(a string) {
+	s.quantity = a
+
+	s.sellOrderSetup(1)
+}
+
+func (s *updateSellOrdersSuite) AliceCreatedTwoSellOrdersEachWithQuantity(a string) {
+	s.quantity = a
+
+	s.sellOrderSetup(2)
+}
+
+func (s *updateSellOrdersSuite) AliceCreatedASellOrderWithAskDenom(a string) {
+	s.askPrice = &sdk.Coin{
+		Denom:  a,
+		Amount: s.askPrice.Amount,
+	}
+
+	s.sellOrderSetup(1)
+}
+
+func (s *updateSellOrdersSuite) AliceCreatedASellOrderWithBatchDenomAndAskDenom(a string, b string) {
+	s.batchDenom = a
+	s.askPrice = &sdk.Coin{
+		Denom:  b,
+		Amount: s.askPrice.Amount,
+	}
+
+	s.sellOrderSetup(1)
+}
+
+func (s *updateSellOrdersSuite) AliceCreatedASellOrderWithTheProperties(a gocuke.DocString) {
+	order := &marketplace.MsgSell_Order{}
+	err := jsonpb.UnmarshalString(a.Content, order)
+	require.NoError(s.t, err)
+
+	s.batchDenom = order.BatchDenom
+	s.quantity = order.Quantity
+	s.askPrice = order.AskPrice
+	s.disableAutoRetire = order.DisableAutoRetire
+	s.expiration = order.Expiration
+
+	s.sellOrderSetup(1)
+}
+
+func (s *updateSellOrdersSuite) AliceCreatedTwoSellOrdersEachWithTheProperties(a gocuke.DocString) {
+	order := &marketplace.MsgSell_Order{}
+	err := jsonpb.UnmarshalString(a.Content, order)
+	require.NoError(s.t, err)
+
+	s.batchDenom = order.BatchDenom
+	s.quantity = order.Quantity
+	s.askPrice = order.AskPrice
+	s.disableAutoRetire = order.DisableAutoRetire
+	s.expiration = order.Expiration
+
+	s.sellOrderSetup(2)
+}
+
+func (s *updateSellOrdersSuite) AliceHasABatchBalanceWithTradableAmountAndEscrowedAmount(a string, b string) {
+	s.aliceTradableAmount = a
+	s.aliceEscrowedAmount = b
+
+	s.aliceBatchBalance()
+}
+
+func (s *updateSellOrdersSuite) AliceAttemptsToUpdateTheSellOrder() {
+	s.res, s.err = s.k.UpdateSellOrders(s.ctx, &marketplace.MsgUpdateSellOrders{
+		Owner: s.alice.String(),
+		Updates: []*marketplace.MsgUpdateSellOrders_Update{
+			{
+				SellOrderId: s.sellOrderId,
+				NewQuantity: s.quantity,
+				NewAskPrice: s.askPrice,
+			},
 		},
 	})
-	assert.ErrorContains(t, err, "ubar is not allowed to be used in sell orders")
+}
+
+func (s *updateSellOrdersSuite) BobAttemptsToUpdateTheSellOrder() {
+	s.res, s.err = s.k.UpdateSellOrders(s.ctx, &marketplace.MsgUpdateSellOrders{
+		Owner: s.bob.String(),
+		Updates: []*marketplace.MsgUpdateSellOrders_Update{
+			{
+				SellOrderId: s.sellOrderId,
+				NewQuantity: s.quantity,
+				NewAskPrice: s.askPrice,
+			},
+		},
+	})
+}
+
+func (s *updateSellOrdersSuite) AliceAttemptsToUpdateTheSellOrderWithId(a string) {
+	id, err := strconv.ParseUint(a, 10, 32)
+	require.NoError(s.t, err)
+
+	s.res, s.err = s.k.UpdateSellOrders(s.ctx, &marketplace.MsgUpdateSellOrders{
+		Owner: s.alice.String(),
+		Updates: []*marketplace.MsgUpdateSellOrders_Update{
+			{
+				SellOrderId: id,
+				NewQuantity: s.quantity,
+				NewAskPrice: s.askPrice,
+			},
+		},
+	})
+}
+
+func (s *updateSellOrdersSuite) AliceAttemptsToUpdateTheSellOrderWithQuantity(a string) {
+	s.res, s.err = s.k.UpdateSellOrders(s.ctx, &marketplace.MsgUpdateSellOrders{
+		Owner: s.alice.String(),
+		Updates: []*marketplace.MsgUpdateSellOrders_Update{
+			{
+				SellOrderId: s.sellOrderId,
+				NewQuantity: a,
+				NewAskPrice: s.askPrice,
+			},
+		},
+	})
+}
+
+func (s *updateSellOrdersSuite) AliceAttemptsToUpdateTheTwoSellOrdersEachWithQuantity(a string) {
+	s.res, s.err = s.k.UpdateSellOrders(s.ctx, &marketplace.MsgUpdateSellOrders{
+		Owner: s.alice.String(),
+		Updates: []*marketplace.MsgUpdateSellOrders_Update{
+			{
+				SellOrderId: 1,
+				NewQuantity: a,
+				NewAskPrice: s.askPrice,
+			},
+			{
+				SellOrderId: 2,
+				NewQuantity: a,
+				NewAskPrice: s.askPrice,
+			},
+		},
+	})
+}
+
+func (s *updateSellOrdersSuite) AliceAttemptsToUpdateTheSellOrderWithAskDenom(a string) {
+	s.res, s.err = s.k.UpdateSellOrders(s.ctx, &marketplace.MsgUpdateSellOrders{
+		Owner: s.alice.String(),
+		Updates: []*marketplace.MsgUpdateSellOrders_Update{
+			{
+				SellOrderId: s.sellOrderId,
+				NewQuantity: s.quantity,
+				NewAskPrice: &sdk.Coin{
+					Denom:  a,
+					Amount: s.askPrice.Amount,
+				},
+			},
+		},
+	})
+}
+
+func (s *updateSellOrdersSuite) AliceAttemptsToUpdateTheSellOrderWithExpiration(a string) {
+	expiration, err := types.ParseDate("expiration", a)
+	require.NoError(s.t, err)
+
+	s.res, s.err = s.k.UpdateSellOrders(s.ctx, &marketplace.MsgUpdateSellOrders{
+		Owner: s.alice.String(),
+		Updates: []*marketplace.MsgUpdateSellOrders_Update{
+			{
+				SellOrderId:   s.sellOrderId,
+				NewQuantity:   s.quantity,
+				NewAskPrice:   s.askPrice,
+				NewExpiration: &expiration,
+			},
+		},
+	})
+}
+
+func (s *updateSellOrdersSuite) AliceAttemptsToUpdateTheSellOrderWithTheProperties(a gocuke.DocString) {
+	update := &marketplace.MsgUpdateSellOrders_Update{}
+	err := jsonpb.UnmarshalString(a.Content, update)
+	require.NoError(s.t, err)
+
+	s.res, s.err = s.k.UpdateSellOrders(s.ctx, &marketplace.MsgUpdateSellOrders{
+		Owner: s.alice.String(),
+		Updates: []*marketplace.MsgUpdateSellOrders_Update{
+			{
+				SellOrderId:       s.sellOrderId,
+				NewQuantity:       update.NewQuantity,
+				NewAskPrice:       update.NewAskPrice,
+				DisableAutoRetire: update.DisableAutoRetire,
+				NewExpiration:     update.NewExpiration,
+			},
+		},
+	})
+}
+
+func (s *updateSellOrdersSuite) AliceAttemptsToUpdateTheTwoSellOrdersEachWithTheProperties(a gocuke.DocString) {
+	update := &marketplace.MsgUpdateSellOrders_Update{}
+	err := jsonpb.UnmarshalString(a.Content, update)
+	require.NoError(s.t, err)
+
+	s.res, s.err = s.k.UpdateSellOrders(s.ctx, &marketplace.MsgUpdateSellOrders{
+		Owner: s.alice.String(),
+		Updates: []*marketplace.MsgUpdateSellOrders_Update{
+			{
+				SellOrderId:       1,
+				NewQuantity:       update.NewQuantity,
+				NewAskPrice:       update.NewAskPrice,
+				DisableAutoRetire: update.DisableAutoRetire,
+				NewExpiration:     update.NewExpiration,
+			},
+			{
+				SellOrderId:       2,
+				NewQuantity:       update.NewQuantity,
+				NewAskPrice:       update.NewAskPrice,
+				DisableAutoRetire: update.DisableAutoRetire,
+				NewExpiration:     update.NewExpiration,
+			},
+		},
+	})
+}
+
+func (s *updateSellOrdersSuite) ExpectNoError() {
+	require.NoError(s.t, s.err)
+}
+
+func (s *updateSellOrdersSuite) ExpectTheError(a string) {
+	require.EqualError(s.t, s.err, a)
+}
+
+func (s *updateSellOrdersSuite) ExpectAliceTradableCreditBalance(a string) {
+	batch, err := s.coreStore.BatchTable().GetByDenom(s.ctx, s.batchDenom)
+	require.NoError(s.t, err)
+
+	balance, err := s.coreStore.BatchBalanceTable().Get(s.ctx, s.alice, batch.Key)
+	require.NoError(s.t, err)
+
+	require.Equal(s.t, a, balance.TradableAmount)
+}
+
+func (s *updateSellOrdersSuite) ExpectAliceEscrowedCreditBalance(a string) {
+	batch, err := s.coreStore.BatchTable().GetByDenom(s.ctx, s.batchDenom)
+	require.NoError(s.t, err)
+
+	balance, err := s.coreStore.BatchBalanceTable().Get(s.ctx, s.alice, batch.Key)
+	require.NoError(s.t, err)
+
+	require.Equal(s.t, a, balance.EscrowedAmount)
+}
+
+func (s *updateSellOrdersSuite) ExpectMarketWithIdAndDenom(a string, b string) {
+	id, err := strconv.ParseUint(a, 10, 32)
+	require.NoError(s.t, err)
+
+	market, err := s.marketStore.MarketTable().Get(s.ctx, id)
+	require.NoError(s.t, err)
+
+	require.Equal(s.t, market.Id, id)
+	require.Equal(s.t, market.CreditTypeAbbrev, s.creditTypeAbbrev)
+	require.Equal(s.t, market.BankDenom, b)
+	require.Equal(s.t, market.PrecisionModifier, uint32(0)) // always zero
+}
+
+func (s *updateSellOrdersSuite) ExpectNoMarketWithId(a string) {
+	id, err := strconv.ParseUint(a, 10, 32)
+	require.NoError(s.t, err)
+
+	_, err = s.marketStore.MarketTable().Get(s.ctx, id)
+	require.ErrorContains(s.t, err, ormerrors.NotFound.Error())
+}
+
+func (s *updateSellOrdersSuite) ExpectSellOrderWithSellerAliceAndTheProperties(a gocuke.DocString) {
+	expected := &marketplace.SellOrder{}
+	err := jsonpb.UnmarshalString(a.Content, expected)
+	require.NoError(s.t, err)
+
+	order, err := s.marketStore.SellOrderTable().Get(s.ctx, expected.Id)
+	require.NoError(s.t, err)
+
+	require.Equal(s.t, expected.Id, order.Id)
+	require.Equal(s.t, s.alice.Bytes(), order.Seller)
+	require.Equal(s.t, expected.AskAmount, order.AskAmount)
+	require.Equal(s.t, expected.Expiration.Seconds, order.Expiration.Seconds)
+	require.Equal(s.t, expected.Expiration.Nanos, order.Expiration.Nanos)
+	require.Equal(s.t, expected.BatchKey, order.BatchKey)
+	require.Equal(s.t, expected.Quantity, order.Quantity)
+	require.Equal(s.t, expected.DisableAutoRetire, order.DisableAutoRetire)
+	require.Equal(s.t, expected.MarketId, order.MarketId)
+	require.Equal(s.t, expected.Maker, order.Maker)
+}
+
+func (s *updateSellOrdersSuite) sellOrderSetup(count int) {
+	totalQuantity := s.quantity
+
+	if count > 1 {
+		c := math.NewDecFromInt64(int64(count))
+		q, err := math.NewDecFromString(s.quantity)
+		require.NoError(s.t, err)
+		t, err := c.Mul(q)
+		require.NoError(s.t, err)
+		totalQuantity = t.String()
+	}
+
+	err := s.coreStore.ClassTable().Insert(s.ctx, &coreapi.Class{
+		Id:               s.classId,
+		CreditTypeAbbrev: s.creditTypeAbbrev,
+	})
+	require.NoError(s.t, err)
+
+	batchKey, err := s.coreStore.BatchTable().InsertReturningID(s.ctx, &coreapi.Batch{
+		Denom: s.batchDenom,
+	})
+	require.NoError(s.t, err)
+
+	err = s.coreStore.BatchBalanceTable().Insert(s.ctx, &coreapi.BatchBalance{
+		BatchKey:       batchKey,
+		Address:        s.alice,
+		EscrowedAmount: totalQuantity,
+	})
+	require.NoError(s.t, err)
+
+	err = s.coreStore.BatchSupplyTable().Insert(s.ctx, &coreapi.BatchSupply{
+		BatchKey:       batchKey,
+		TradableAmount: totalQuantity,
+	})
+	require.NoError(s.t, err)
+
+	marketKey, err := s.marketStore.MarketTable().InsertReturningID(s.ctx, &api.Market{
+		CreditTypeAbbrev: s.creditTypeAbbrev,
+		BankDenom:        s.askPrice.Denom,
+	})
+	require.NoError(s.t, err)
+
+	order := &api.SellOrder{
+		Seller:            s.alice,
+		BatchKey:          batchKey,
+		Quantity:          s.quantity,
+		MarketId:          marketKey,
+		AskAmount:         s.askPrice.Amount.String(),
+		DisableAutoRetire: s.disableAutoRetire,
+	}
+
+	if s.expiration != nil {
+		order.Expiration = timestamppb.New(*s.expiration)
+	}
+
+	sellOrderId, err := s.marketStore.SellOrderTable().InsertReturningID(s.ctx, order)
+	require.NoError(s.t, err)
+	require.Equal(s.t, sellOrderId, s.sellOrderId)
+
+	for i := 1; i < count; i++ {
+		order.Id = 0 // reset sell order id
+		err = s.marketStore.SellOrderTable().Insert(s.ctx, order)
+		require.NoError(s.t, err)
+	}
+}
+
+func (s *updateSellOrdersSuite) aliceBatchBalance() {
+	batch, err := s.coreStore.BatchTable().GetByDenom(s.ctx, s.batchDenom)
+	if err == ormerrors.NotFound {
+		classKey, err := s.coreStore.ClassTable().InsertReturningID(s.ctx, &coreapi.Class{
+			Id:               s.classId,
+			CreditTypeAbbrev: s.creditTypeAbbrev,
+		})
+		require.NoError(s.t, err)
+
+		projectKey, err := s.coreStore.ProjectTable().InsertReturningID(s.ctx, &coreapi.Project{
+			ClassKey: classKey,
+		})
+		require.NoError(s.t, err)
+
+		batchKey, err := s.coreStore.BatchTable().InsertReturningID(s.ctx, &coreapi.Batch{
+			ProjectKey: projectKey,
+			Denom:      s.batchDenom,
+		})
+		require.NoError(s.t, err)
+
+		batch.Key = batchKey
+	} else if err != nil {
+		require.NoError(s.t, err)
+	}
+
+	// Save because batch balance may already exist from sell order setup
+	err = s.coreStore.BatchBalanceTable().Save(s.ctx, &coreapi.BatchBalance{
+		BatchKey:       batch.Key,
+		Address:        s.alice,
+		TradableAmount: s.aliceTradableAmount,
+		EscrowedAmount: s.aliceEscrowedAmount,
+	})
+	require.NoError(s.t, err)
 }
